# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for Najammelha, a Kuwaiti community reporting platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  isAdmin: boolean;
  points: number;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}

export interface LogoutResponse {
  success: boolean;
}

export interface Category {
  id: string;
  name: string;
  label: string;
  icon: string;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  received: 'received',
  reviewing: 'reviewing',
  referred: 'referred',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export type ReportAgencySource = typeof ReportAgencySource[keyof typeof ReportAgencySource] | null;


export const ReportAgencySource = {
  rules: 'rules',
  ai: 'ai',
  admin: 'admin',
} as const;

export interface Report {
  id: number;
  userId: string;
  authorName?: string;
  image: string;
  title: string;
  description: string;
  category: string;
  categoryLabel: string;
  latitude: number;
  longitude: number;
  locationName: string;
  status: ReportStatus;
  statusLabel: string;
  points: number;
  supportCount: number;
  createdAt: string;
  updatedAt: string;
  isDemo: boolean;
  /** True when the coordinates came from the reporter's device GPS. */
  locationExact?: boolean;
  agencyId?: string | null;
  agencyName?: string | null;
  /** Why the router suggested this agency (admin only). */
  agencyReason?: string | null;
  agencyConfidence?: number | null;
  agencySource?: ReportAgencySource;
}

export interface ReportInput {
  /** @minLength 1 */
  image: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 2
     * @maxLength 800
     */
  description: string;
  category: string;
  /**
     * @minimum 28
     * @maximum 31
     */
  latitude: number;
  /**
     * @minimum 46
     * @maximum 49
     */
  longitude: number;
  /**
     * @minLength 2
     * @maxLength 120
     */
  locationName: string;
  locationExact?: boolean;
}

export type ReportUpdateStatus = typeof ReportUpdateStatus[keyof typeof ReportUpdateStatus];


export const ReportUpdateStatus = {
  received: 'received',
  reviewing: 'reviewing',
  referred: 'referred',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export interface ReportUpdate {
  status?: ReportUpdateStatus;
  /** @minimum 0 */
  supportCount?: number;
  /** Admin override of the suggested agency. */
  agencyId?: string;
}

export interface PlatformStats {
  totalReports: number;
  resolvedReports: number;
  communityContributions: number;
  activeAreas: number;
}

export interface LevelInfo {
  id: string;
  name: string;
  minPoints: number;
  reportPoints: number;
  resolvedBonus: number;
  perks: string[];
}

export interface Notification {
  id: number;
  title: string;
  body: string;
  createdAt: string;
  read: boolean;
}

export interface UserDashboard {
  name: string;
  initials: string;
  points: number;
  lifetimePoints: number;
  reportsCount: number;
  resolvedCount: number;
  contributionRate: number;
  level: string;
  levelId: string;
  nextLevelPoints: number;
  levels: LevelInfo[];
  reports: Report[];
  notifications: Notification[];
}

export interface BreakdownItem {
  label: string;
  value: number;
}

export interface AdminOverview {
  totalReports: number;
  newReports: number;
  reviewingReports: number;
  resolvedReports: number;
  usersCount: number;
  totalPoints: number;
  totalRedemptions: number;
  totalPointsRedeemed: number;
  byCategory: BreakdownItem[];
  byArea: BreakdownItem[];
}

export interface Reward {
  id: number;
  title: string;
  description: string;
  image: string;
  partnerName: string;
  discountLabel: string;
  costPoints: number;
  stock: number | null;
  isActive: boolean;
  createdAt: string;
}

export interface RewardInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 2
     * @maxLength 800
     */
  description: string;
  /** @minLength 1 */
  image: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  partnerName: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  discountLabel: string;
  /** @minimum 1 */
  costPoints: number;
  /** @minimum 0 */
  stock?: number | null;
}

export interface RewardUpdate {
  /**
     * @minLength 2
     * @maxLength 120
     */
  title?: string;
  /**
     * @minLength 2
     * @maxLength 800
     */
  description?: string;
  /** @minLength 1 */
  image?: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  partnerName?: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  discountLabel?: string;
  /** @minimum 1 */
  costPoints?: number;
  /** @minimum 0 */
  stock?: number | null;
  isActive?: boolean;
}

export interface Redemption {
  id: number;
  rewardId: number;
  code: string;
  costPoints: number;
  rewardTitle: string;
  partnerName: string;
  discountLabel: string;
  createdAt: string;
}

export interface ImageUpload {
  file: Blob;
}

export interface ImageObject {
  objectPath: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListReportsParams = {
status?: ListReportsStatus;
category?: string;
sort?: ListReportsSort;
search?: string;
mine?: boolean;
};

export type ListReportsStatus = typeof ListReportsStatus[keyof typeof ListReportsStatus];


export const ListReportsStatus = {
  all: 'all',
  received: 'received',
  reviewing: 'reviewing',
  referred: 'referred',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export type ListReportsSort = typeof ListReportsSort[keyof typeof ListReportsSort];


export const ListReportsSort = {
  latest: 'latest',
  supported: 'supported',
} as const;