# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for Najammelha, a Kuwaiti community reporting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List public reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, received, reviewing, referred, resolved, closed]
            default: all
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, supported]
            default: latest
        - name: search
          in: query
          schema:
            type: string
        - name: mine
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a new community report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Validation error
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report by id
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update a report status or support count
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report (admin)
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "204":
          description: Report deleted
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /reports/{id}/support:
    post:
      operationId: supportReport
      tags: [reports]
      summary: Support a report once per user
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: Supported report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
...
  /storage/upload:
    post:
      tags: [Storage]
      operationId: uploadImage
      summary: Upload an image to blob storage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageUpload"
      responses:
        "200":
          description: Uploaded image path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageObject"
        "400":
          description: Invalid image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/objects/{key}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded image
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema: { type: string, format: binary }
        "404":
          description: Object not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    ReportId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Report:
      type: object
      required: [id, userId, image, title, description, category, categoryLabel, latitude, longitude, locationName, status, statusLabel, points, supportCount, createdAt, updatedAt, isDemo]
      properties:
        id: { type: integer }
        userId: { type: string }
        authorName: { type: string }
        image: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        categoryLabel: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        locationName: { type: string }
        status:
          type: string
          enum: [received, reviewing, referred, resolved, closed]
        statusLabel: { type: string }
        points: { type: integer }
        supportCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        isDemo: { type: boolean }
        locationExact:
          type: boolean
          description: True when the coordinates came from the reporter's device GPS.
        agencyId: { type: string, nullable: true }
        agencyName: { type: string, nullable: true }
        agencyReason:
          type: string
          nullable: true
          description: Why the router suggested this agency (admin only).
        agencyConfidence: { type: number, nullable: true }
        agencySource:
          type: string
          nullable: true
          enum: [rules, ai, admin]
    ReportInput:
      type: object
      required: [image, title, description, category, latitude, longitude, locationName]
      properties:
        image: { type: string, minLength: 1 }
        title: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, minLength: 2, maxLength: 800 }
        category: { type: string }
        latitude: { type: number, minimum: 28, maximum: 31 }
        longitude: { type: number, minimum: 46, maximum: 49 }
        locationName: { type: string, minLength: 2, maxLength: 120 }
        locationExact: { type: boolean }
    ReportUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [received, reviewing, referred, resolved, closed]
        supportCount:
          type: integer
          minimum: 0
        agencyId:
          type: string
          description: Admin override of the suggested agency.
...
    ImageUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    ImageObject:
      type: object
      required: [objectPath]
      properties:
        objectPath: { type: string }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }