# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for Najammelha, a Kuwaiti community reporting platform
 * OpenAPI spec version: 0.1.0
 */
import type { ReportAgencySource } from './reportAgencySource';
import type { ReportStatus } from './reportStatus';

export interface Report {
  id: number;
  userId: string;
  authorName?: string;
  image: string;
  title: string;
  description: string;
  category: string;
  categoryLabel: string;
  latitude: number;
  longitude: number;
  locationName: string;
  status: ReportStatus;
  statusLabel: string;
  points: number;
  supportCount: number;
  createdAt: Date;
  updatedAt: Date;
  isDemo: boolean;
  /** True when the coordinates came from the reporter's device GPS. */
  locationExact?: boolean;
  agencyId?: string | null;
  agencyName?: string | null;
  /** Why the router suggested this agency (admin only). */
  agencyReason?: string | null;
  agencyConfidence?: number | null;
  agencySource?: ReportAgencySource;
}