# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for Najammelha, a Kuwaiti community reporting platform
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  /** @minLength 1 */
  image: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 2
     * @maxLength 800
     */
  description: string;
  category: string;
  /**
     * @minimum 28
     * @maximum 31
     */
  latitude: number;
  /**
     * @minimum 46
     * @maximum 49
     */
  longitude: number;
  /**
     * @minLength 2
     * @maxLength 120
     */
  locationName: string;
  locationExact?: boolean;
}